- Implement AuthService, UserService - Extend Person with .interests: [DiscussionTopic] - Make MatchState dependable to UserService with no extra parameters in calls

+++ CoffeeBreak/CoffeeBreak/Services/MatchService.swift
@@ -6,13 +6,14 @@
 //
 
 import Foundation
+import Combine
 
 // MARK: - IMatchService
 
 protocol IMatchService: ObservableObject {
     var matchState: MatchState { get }
     var readyForRequests: Bool { get }
-    func requestCoffeeBreak(interests: [DiscussionTopic])
+    func requestCoffeeBreak()
     func cancelCoffeeBreakRequest()
 }
 
@@ -38,20 +39,30 @@ enum MatchState: Equatable {
 
 // MARK: - MatchService Implementation
 
-final class MatchService<NS: INetworkService>: ObservableObject, IMatchService {
+final class MatchService<NS: INetworkService, US: IUserService>: ObservableObject, IMatchService {
     // State
-    @Published private(set) var matchState: MatchState = .uninitiated
-    var readyForRequests: Bool { return matchState == .idle }
+    @Published private(set) var matchState: MatchState = .uninitiated {
+        didSet { print("Match service: \(matchState)") }
+    }
+    var readyForRequests: Bool { return matchState == .idle && currentUser != nil }
     private var currentUserQueuePosition: MeetupQueueElement.ID?
     // Dependencies
     private let networkService: NS
-    private var currentUser = Person.dummy
+    private var currentUser: Person?
+    private let userService: US
+    // Etc
     private var queueSubscription: INetworkService.NetworkServiceSubscription?
     private var loungeRoomsSubscription: INetworkService.NetworkServiceSubscription?
+    private var bag = Set<AnyCancellable>()
 
-    init(networkService: NS) {
+    init(networkService: NS, userService: US) {
         self.networkService = networkService
-        subscribeToLoungeRoomUpdates()
+        self.userService = userService
+        // Required to trigger Firebase after App's didFinishLaunching / FB configuration
+        DispatchQueue.main.async { [weak self] in
+            self?.subscribeToLoungeRoomUpdates()
+            self?.subscribeToUserService()
+        }
     }
 
     deinit {
@@ -60,9 +71,9 @@ final class MatchService<NS: INetworkService>: ObservableObject, IMatchService {
         unsubscribeFromQueueUpdates()
     }
 
-    func requestCoffeeBreak(interests: [DiscussionTopic]) {
+    func requestCoffeeBreak() {
         // Avoids requesting unless in idle mode
-        guard matchState == .idle else { return }
+        guard readyForRequests, let currentUser = currentUser else { return }
         // Sets internal state
         matchState = .searching
         // Adds request to the server
@@ -70,11 +81,11 @@ final class MatchService<NS: INetworkService>: ObservableObject, IMatchService {
             id: UUID().uuidString,
             timeCreated: .now,
             timeExpires: .now + Constants.queueElementExpiration,
-            topicIds: interests.map({ $0.rawValue }),
+            topicIds: currentUser.interests.map({ $0.rawValue }),
             userId: currentUser.id)
         )
         // Subscribes to network service meetup queue updates
-        subscribeToQueueUpdates(ownInterests: interests)
+        subscribeToQueueUpdates(ownInterests: currentUser.interests)
     }
     
     func cancelCoffeeBreakRequest() {
@@ -117,15 +128,28 @@ private extension MatchService {
             loungeRoomsSubscription = nil
         }
     }
+
+    func subscribeToUserService() {
+        userService.currentUser.publisher.sink { [weak self] person in
+            self?.cancelCoffeeBreakRequest()
+            self?.unsubscribeFromLoungeRoomUpdates()
+            self?.unsubscribeFromQueueUpdates()
+            self?.matchState = .uninitiated
+            self?.currentUser = person
+            self?.subscribeToLoungeRoomUpdates()
+        }
+        .store(in: &bag)
+    }
 }
 
 // MARK: - Processing updates
 
 private extension MatchService {
     func processLoungeRooms(_ rooms: [LoungeRoom]) {
+        guard let currentUser = currentUser else { return }
         for room in rooms {
             if room.timeExpires < .now { networkService.removeLoungeRoom(room.id); continue }
-            if !room.members.contains(self.currentUser.id) { continue }
+            if !room.members.contains(currentUser.id) { continue }
             // Present in a room at this point
             guard let other = room.members.first(where: { $0 != currentUser.id }) else { return }
             matchState = .match(with: other)
@@ -139,8 +163,8 @@ private extension MatchService {
     func processMeetupQueue(_ queue: [MeetupQueueElement], ownInterests: [DiscussionTopic]) {
         // Removes expired items
         removeExpiredItems(queue)
-        // Avoids proceeding unless in searching mode
-        guard matchState == .searching else { return }
+        // Avoids proceeding unless in searching mode / unauthed
+        guard matchState == .searching, let currentUser = currentUser else { return }
         // Looks for a match
         let ownInterests = Set(ownInterests.map({ $0.rawValue }))
         let sortedQueue = queue.sorted(by: { $0.timeCreated <= $1.timeCreated })
